management-system-backend/auth/requirements.txt
django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2==2.9.9
python-dotenv